Move and split RippleLikeOperationQuery into files.
