+++ ledger-core-ripple/src/ripple/RippleLikeOperation.cpp
@@ -53,43 +53,5 @@ namespace ledger {
               getOperationType()
           );
         }
-
-        RippleLikeOperationQuery::RippleLikeOperationQuery(
-            const std::shared_ptr<api::QueryFilter>& headFilter,
-            const std::shared_ptr<DatabaseSessionPool>& pool,
-            const std::shared_ptr<api::ExecutionContext>& context,
-            const std::shared_ptr<api::ExecutionContext>& mainContext
-        ): OperationQuery(headFilter, pool, context, mainContext) {
-        }
-
-
-        std::shared_ptr<RippleLikeOperation> RippleLikeOperationQuery::createOperation(
-            std::shared_ptr<AbstractAccount> &account
-        ) {
-            RippleLikeBlockchainExplorerTransaction tx;
-            return std::make_shared<RippleLikeOperation>(
-                std::make_shared<RippleLikeBlockchainExplorerTransaction>(tx),
-                account->getWallet()->getCurrency()
-            );
-        }
-
-        void RippleLikeOperationQuery::inflateCompleteTransaction(
-            soci::session& sql,
-            std::string const& accountUid,
-            RippleLikeOperation& operation
-        ) {
-            std::string transactionHash;
-            RippleLikeBlockchainExplorerTransaction tx;
-
-            sql << "SELECT transaction_hash FROM ripple_operations WHERE uid = :uid", soci::use(operation.uid), soci::into(transactionHash);
-            RippleLikeTransactionDatabaseHelper::getTransactionByHash(
-                sql,
-                transactionHash,
-                tx
-            );
-
-            auto currency = _accounts[accountUid]->getWallet()->getCurrency();
-            operation = RippleLikeOperation(std::make_shared<RippleLikeBlockchainExplorerTransaction>(tx), currency);
-        }
     }
 }
